+++ etc/example-setup/config/config.mjs
@@ -161,11 +161,12 @@ const applications = [
     ]
   },
   {
-    name:  'myHosts',
-    class: HostRouter,
+    name:       'myHosts',
+    class:      HostRouter,
+    ignoreCase: true,
     hosts: {
       '*':         'myPaths',
-      '127.0.0.1': 'mySeries'
+      '127.0.0.1': 'mySeries',
     }
   },
   {