+++ src/webapp-builtins/tests/HostRouter.test.js
@@ -17,6 +17,13 @@ describe('constructor', () => {
     })).not.toThrow();
   });
 
+  test('accepts a valid minimal configuration with `ignoreCase`', () => {
+    expect(() => new HostRouter({
+      hosts:      {},
+      ignoreCase: false
+    })).not.toThrow();
+  });
+
   test('accepts a valid configuration with several non-wildcard hosts', () => {
     expect(() => new HostRouter({
       hosts: {
@@ -49,6 +56,42 @@ describe('constructor', () => {
         '10.0.0.123':    'app2',
         '0:0::1':        'app3',
         '1234:abcd::98': 'app3'
+      }
+    })).not.toThrow();
+  });
+
+  test('does not allow two names that differ only in case when `ignoreCase === true`', () => {
+    expect(() => new HostRouter({
+      ignoreCase: true,
+      hosts: {
+        'Boop.bop': 'app1',
+        'booP.bop': 'app2'
+      }
+    })).toThrow();
+
+    expect(() => new HostRouter({
+      ignoreCase: true,
+      hosts: {
+        '*.ZONK': 'app1',
+        '*.ZoNK': 'app2'
+      }
+    })).toThrow();
+  });
+
+  test('allows two names that differ only in case when `ignoreCase === false`', () => {
+    expect(() => new HostRouter({
+      ignoreCase: false,
+      hosts: {
+        'Boop.bop': 'app1',
+        'booP.bop': 'app2'
+      }
+    })).not.toThrow();
+
+    expect(() => new HostRouter({
+      ignoreCase: false,
+      hosts: {
+        '*.ZONK': 'app1',
+        '*.ZoNK': 'app2'
       }
     })).not.toThrow();
   });
@@ -223,6 +266,37 @@ describe('_impl_handleRequest()', () => {
     );
 
     await expectApp(hr, 'zorch.splat', 'mockApp1');
+  });
+
+  test('routes to a case-folded DNS name when `ignoreCase === true`', async () => {
+    const hr = await makeInstance(
+      {
+        ignoreCase: true,
+        hosts: {
+          'splat':             'mockApp2',
+          'ZORCH.splat':       'mockApp1',
+          'blurp.zorch.splat': 'mockApp2'
+        }
+      },
+      { appCount: 2 }
+    );
+
+    await expectApp(hr, 'zorch.SPLAT', 'mockApp1');
+  });
+
+  test('routes to a case-matched DNS name when `ignoreCase === false`', async () => {
+    const hr = await makeInstance(
+      {
+        ignoreCase: false,
+        hosts: {
+          'ZORCH.splat': 'mockApp1',
+          'zorch.SPLAT': 'mockApp2'
+        }
+      },
+      { appCount: 2 }
+    );
+
+    await expectApp(hr, 'zorch.SPLAT', 'mockApp2');
   });
 
   test('does not route to an exact-match DNS name as if it were a wildcard', async () => {